move qt json widget builder to _vendor folder.  switch to qtpy
