--- napari/_qt/qt_json_builder/test_schema3.json
@@ -1,1 +1,0 @@
-{"title": "Application settings", "description": "Main application settings.", "type": "object", "properties": {"qt_binding": {"title": "Qt Binding", "default": "pyside2", "env_names": ["napari_settings_application_qt_binding"], "type": "string"}, "highlight_thickness": {"title": "Highlight Thickness", "default": 1, "env_names": ["napari_settings_application_highlight_thickness"], "type": "integer"}, "opt_in_telemetry": {"title": "Opt In Telemetry", "description": "Check to enable telemtry measurements", "default": false, "env_names": ["napari_settings_application_opt_in_telemetry"], "type": "boolean"}, "font_plain_family": {"title": "Font Plain Family", "env_names": ["napari_settings_application_font_plain_family"], "type": "string"}, "font_plain_size": {"title": "Font Plain Size", "env_names": ["napari_settings_application_font_plain_size"], "type": "integer"}, "font_rich_family": {"title": "Font Rich Family", "env_names": ["napari_settings_application_font_rich_family"], "type": "string"}, "font_rich_size": {"title": "Font Rich Size", "env_names": ["napari_settings_application_font_rich_size"], "type": "integer"}, "plugins_call_order": {"title": "Plugins Call Order", "default": [], "env_names": ["napari_settings_application_plugins_call_order"], "type": "array", "items": {"type": "string"}}, "version": {"title": "Version", "default": [0, 1, 0], "env_names": ["napari_settings_application_version"], "type": "array", "items": {}}}}
