+++ napari/_qt/qt_main_window.py
@@ -33,8 +33,8 @@ from .dialogs.qt_plugin_dialog import QtPluginDialog
 from .dialogs.qt_plugin_report import QtPluginErrReporter
 from .dialogs.screenshot_dialog import ScreenshotDialog
 from .perf.qt_debug_menu import DebugMenu
+from .preferences_schema import PreferencesDialog
 from .qt_event_loop import NAPARI_ICON_PATH, get_app
-from .qt_json_builder.preferences_schema import PreferencesDialog
 from .qt_resources import get_stylesheet
 from .qt_viewer import QtViewer
 from .utils import QImg2array
