+++ napari/_vendor/qt_json_builder/LICENSE
