+++ napari/_vendor/qt_json_builder/qt_jsonschema_form/signal.py
